requirements.txt
virtualenv
click==8.1.3
Flask==2.2.3
Flask-Reuploaded==1.3.0
Flask-Uploads==0.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3

# LaMa requirements
torch
torchvision
pyyaml
tqdm
numpy
easydict==1.9.0
scikit-image
scikit-learn==0.24.2
opencv-python
tensorflow
joblib
matplotlib
pandas
albumentations==0.5.2
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
kornia==0.5.0
webdataset
packaging
scikit-learn==0.24.2
